requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
jwt==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
